Refactor playback methods to use Redux thunks for track and playlist play
